feat: cr tdl models declared in separate files
